Updated to make sure it is grabbing correct date ranges
couple of details. First, the year that you put in the url gives you the gamelog for the games that were in the season that ended in that year. So, if you wanted to find games from October to December 2018, the year would be marked as 2018, and only games from the 2017-2018 season would be fetched. Further, there was no way to 'prune out' games that weren't in the daterange, so I just added the .loc to filter out the games based on the string